WIP - Add button to share event invitation with another user's cozy
This is just the visual part: a user can click on the button in the
shape of a cloud to express the fact (s)he wants to share the event
directly with the cozy of the guest. There is a visual feedback but
nothing behind the scene is yet implemented.

For this behavior to operate several hypotheses are made: in the contact
information of the guest, there must be a field in the "other
information" part entitled "COZY" that contains the url of the guest's
cozy.