WIP - first iteration of sharing
It is now possible to share an event from the calendar to another cozy
provided a few conditions:

* there must be in the other information section of the targeted contact
a field named "COZY" (all caps) renamed from the "other" category. This
field must have the url of the target's cozy as value
(e.g. http://192.168.50.6:9104)

* the target's cozy and the cozy initiating the sharing process must run,
for now, the forks of the: cozy-data-system, cozy-home, and cozy-proxy
hosted on Gara64's Github account on their branch sharing

Signed-off-by: Loudet Julien <julien.loudet@protonmail.com>